spectator bet

+++ HorseRacingTournamentManagementSystem_0/Modules/Users/DTOs/UserResponseDto.cs
@@ -12,5 +12,5 @@ public class UserResponseDto
     public string? Avatar { get; set; }
     public int? ExperienceYear { get; set; }
     public int? ExpYears { get; set; }
-    public int? TotalPoints { get; set; }
+    public double? TotalPoints { get; set; }
 }
